Separate out bugreport functions
